Names inferred in deconstruction affect refactorings

+++ src/EditorFeatures/CSharpTest/CodeActions/InlineTemporary/InlineTemporaryTests.cs
@@ -4261,7 +4261,7 @@ class C
     static int y = 1;
     void M()
     {
-        var t = ((C.y, (C.y, _)) = (1, (C.y, 3)));
+        var t = (((int)C.y, ((int)C.y, _)) = (1, (C.y, 3)));
     }
 }";
             await TestInRegularAndScriptAsync(code, expected, ignoreTrivia: false);
@@ -4287,7 +4287,7 @@ class C
     static int y = 1;
     void M()
     {
-        var t = ((C.y, _) = (1, 2));
+        var t = (((int)C.y, _) = (1, 2));
     }
 }";
             await TestInRegularAndScriptAsync(code, expected, ignoreTrivia: false);
